Uses CUDA as default backend if none selected
If no backend (CUDA, OpenMP, HIP) is explicitly selected, this commit configures the project to use CUDA as the default backend.

This ensures that the library can be built and used out-of-the-box without requiring users to manually specify a backend when CUDA is desired. It also adds "cuda" as a dependency in vcpkg.json.