Fix error in x_vcpkg_install_local_dependencies (#46061)
OBJECT_LIBARIES must be ignored in x_vcpkg_install_local_dependencies,
because the generator expression $<TARGET_FILE_NAME:tgt> is invalid for
them.

Check whether the target's type is in a list of supported types and
ignore it if it's not. Should CMake introduce a new target type that
doesn't support $<TARGET_FILE_NAME:tgt> then this code won't produce
an error.