Do not check version of system ninja on FreeBSD (microsoft/vcpkg#46344)
VCPKG upgraded it's ninja version to 1.12.1 in microsoft/vcpkg@f576077 / microsoft/vcpkg#41980 .
FreeBSD had ninja-1.12.1 in it's third party package repository, but reverted it to 1.11.0 (freebsd/freebsd-ports@1e1a1e9), as issues with ccache were discovered when using ninja-1.12.1.

vcpkg does not deploy ninja on FreeBSD, and does not accept the version found on the system, either.

Remove the version check on FreeBSD's branch, to allow vcpkg to use the pkg/ports version even if it is "outdated".