[fontconfig] Hard-depend on libintl (microsoft/vcpkg#46523)
meson is passing libintl to the linker even if NLS is disabled. On systems
with an ABI incompatible libintl installed (i.e. outside of vcpkg's
deployment), this will cause the build to fail.

While on the Azure CI only the osx build agents are failing, libintl is
commonly installed on end user systems, though in most systems it's ABI
compatible to the vcpkg code.

Always use vcpkg's libintl in order to prevent trouble.