[faudio] add native default feature for Windows
Add the `native` feature in `default-features` to use the
`PLATFORM_WIN32` backend for Windows targets, which was the default
behavior previously and continues to be.

This allows installing `faudio[core]` to build with the SDL3 backend,
which is useful for Windows XP which does not support the Windows
platform backend.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>