feat(libpq): Update tzcode to 2025b
This commit updates the tzcode MSYS2 package, a build-time dependency, from version 2023c-1 to 2025b-1.

The port-version has been incremented to reflect this change.
This change was added because the msys2 repo removed the existing version of tzcode (2023c). Currently `libpq` builds fail because the file cannot be downloaded (404 not found).