[qt5-webengine] Fix python application name (microsoft/vcpkg#47010)
In commit 358f389 the default python distribution that is automatically
deployed by vcpkg contains a python application called `python.exe`.
Unlike some predecessors, a `python3.exe` is not deployed. The portfile
of qt5-webengine, however, is still instructing qmake that `python3`
is the name of the python3 interpreter.

Fix qmake failing to discover python by adapting the application name.