Disable some very expensive ports on x64-osx.
https://dev.azure.com/vcpkg/public/_build/results?buildId=120023

We are no longer able to finish an x64-osx run within the 48 hour time limit, and x64-osx is on its way out. In particular, as of upcoming macOS 26 we will be unable to run a current version of macOS on any hardware available to the vcpkg team.

Given that the hardware has become unobtanium and we see no practical way to fix the performance issues, I'm disabling particularly expensive ports with relatively few downstream dependencies.

The following runs took >39 minutes on x64-osx:

```console
2025-09-14T16:31:00.7342830Z Elapsed time to handle gz-sim:x64-osx: 4 h
2025-09-14T02:34:36.3603110Z Elapsed time to handle llvm:x64-osx: 3.2 h
2025-09-13T19:15:46.8911760Z Elapsed time to handle vtk:x64-osx: 2.1 h
2025-09-15T01:44:23.0634700Z Elapsed time to handle libtorch:x64-osx: 1.7 h
2025-09-13T15:16:32.5876700Z Elapsed time to handle qtdeclarative:x64-osx: 1.5 h
2025-09-15T03:56:41.6201490Z Elapsed time to handle mdl-sdk:x64-osx: 59 min
2025-09-14T04:06:05.9900930Z Elapsed time to handle osg:x64-osx: 53 min
2025-09-13T04:58:45.6678750Z Elapsed time to handle openmpi:x64-osx: 44 min
2025-09-13T08:47:40.4548680Z Elapsed time to handle arrayfire:x64-osx: 39 min
```

* I'm assuming that the gz-* ports should all be implicated by gz-sim.
    * Also implicates sdformat
* llvm,vtk,openmpi,qtdeclarative have too many downstream dependents to disable.
* osg also implicates:
    * dartsim[gui-osg]
    * osg-qt (also depends on no longer in support qt5)
    * osgearth
* arrayfire also implicates:
    * flashlight-cpu (already skipped)
    * flashlight-cuda (already skipped)