[vulkan-validationlayers] disable use of vendored parallel_hashmap
vcpkg guidelines require that ports have exactly one dependency provider. the upstream source here attempts to autodetect a system-provided parallel_hashmap library, and builds its own if one is not present. while convenient for upstream, this results in different sources being available. in this diff we patch the upstream source to remove the vendored copy and streamline the cmake logic for finding it.