[node-api-headers] Fix node-api-headers CMakeLists.txt
Add a CMake minimum required version of 3.0 to the node-api-headers port in order to fix builds on CMake >= 4.0 on Windows.

When compiled on Windows, this port installs CMakeLists.txt to the root of the node-api-headers repository before building the port. Up until now, the CMakeLists.txt file neglected to declare a minimum CMake version even though it is required (Refer to CMake policy CMP0000).

The OLD behavior of CMP0000, which was to silently ignore the missing invocation of cmake_minimum_required, was removed in CMake version 4.0. CMake 4.0+ will fail to configure and build projects which do not specify a minimum CMake version.

Bump port-version.