[skia] remove pathediting feature and always expose mutable SkPath API
In skia 143, SkPath has started to migrate to have immutable geometry.

The mutable methods on SkPath are hidden when SK_HIDE_PATH_EDIT_METHODS
is defined, and 832f6e5 updated the GN build to always define it. Since
the functionality is not yet removed from skia, instead of adding a
feature that does not correspond to any upstream build config, we
instead add a patch that removes the SK_HIDE_PATH_EDIT_METHODS so
consumers of this vcpkg port version can still use the mutable SkPath
API.