[aom] Fix building with 3.x NASM
Newer NASM versions have a different help text format which caused `test_nasm` to fail on x86 builds, which results in the build failing.

This fix is cherry-picked from upstream on master (https://aomedia-review.googlesource.com/c/aom/+/203481), but there's no new release tagged with the fix in yet, hence the need for the patch.