[shaderc] Fix pkgconfig file generation
In our port, we apply a patch to the build to link against the `glslang` and `SPIRV-Tools` libraries.
While the generated cmake config files reflect this change, the generated pkgconfig file did not. This change fixes that.

Additionally, I also added in creation of a `shaderc` pkgconfig file in addition to the `shaderc_static` pkgconfig file when building statically.
This change brings it in-line with how pkgconfig files are generated for other libraries (the static library has the same name as the shared library),
and fixes scenarios where we may be building a static `shaderc` library but dynamic other libraries that depend on it.